+++ src/main/java/utils/ExtensionType.java
@@ -1,4 +1,4 @@
-package webserver;
+package utils;
 
 import java.io.File;
 import java.util.Arrays;
