+++ src/main/java/webserver/handler/ErrorHandler.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.handler;
 
 public class ErrorHandler {
     private byte[] errorBody;
